add custom exception

+++ controller.py
@@ -1,21 +1,29 @@
 import view
 import text
+import error
+import utils
 from model.phonebook import PhoneBook
 
 
 class MainController:
+    """Основной обработчик"""
     def __init__(self, file_path):
         self.phonebook = PhoneBook(file_path)
+        self.is_running = True
 
     def open_phonebook(self):
-        self.phonebook.load_json()
-        view.show_message(text.phonebook_load_successful)
-
+        try:
+            self.phonebook.load_json()
+        except FileNotFoundError as e:
+            view.show_message(text.error.format(error=e))
+        except error.InvalidJSONError as e:
+            view.show_message(text.error.format(error=e))
+        else:
+            view.show_message(text.phonebook_load_successful)
 
     def save_phonebook(self):
         self.phonebook.save_json()
         view.show_message(text.phonebook_save_successful)
-
 
     def show_contacts(self):
         view.show_contacts(self.phonebook.contacts, text.empty_phonebook_error)
@@ -26,16 +34,18 @@ class MainController:
         found_contacts = self.phonebook.find_contact(find_word)
         view.show_contacts(found_contacts, text.no_found_contact.format(word=find_word))
 
-
     def create_contact(self):
         data = view.input_date([
             text.input_contact_name,
             text.input_contact_phone,
             text.input_contact_comment,
         ])
-        self.phonebook.create_contact(data)
-        view.show_message(text.contact_created_successful.format(name=data[0]))
-
+        try:
+            self.phonebook.create_contact(data)
+        except error.InvalidPhoneError as e:
+            view.show_message(text.error.format(error=e))
+        else:
+            view.show_message(text.contact_created_successful.format(name=data[0]))
 
     def edit_contact(self):
         contact_id = view.input_date(text.input_id_contact)
@@ -49,9 +59,14 @@ class MainController:
             'phone': new_data[1],
             'comment': new_data[2],
         }
-        edited_contact = self.phonebook.edit_contact(contact_id, new_data)
-        view.show_message(text.contact_edited_successful.format(name=edited_contact['name']))
-
+        try:
+            edited_contact = self.phonebook.edit_contact(contact_id, new_data)
+        except error.InvalidContactIDError as e:
+            view.show_message(text.error.format(error=e))
+        except error.ContactNotFoundError as e:
+            view.show_message(text.error.format(error=e))
+        else:
+            view.show_message(text.contact_edited_successful.format(name=edited_contact['name']))
 
     def delete_contact(self):
         contact_id = view.input_date(text.input_id_contact)
@@ -59,14 +74,22 @@ class MainController:
             view.show_message(text.input_id_contact_error)
             return
         contact_id = int(contact_id)
-        contact = self.phonebook.delete_contact(contact_id)
-        view.show_message(text.contact_deleted_successful.format(name=contact.name))
 
+        user_confirm = utils.ask_yes_no(text.confirm_delete_contact, default='no')
+        if not user_confirm:
+            view.show_message(text.cancel_operation)
+            return
+        try:
+            contact = self.phonebook.delete_contact(contact_id)
+        except error.InvalidContactIDError as e:
+            view.show_message(text.error.format(error=e))
+        except error.ContactNotFoundError as e:
+            view.show_message(text.error.format(error=e))
+        else:
+            view.show_message(text.contact_deleted_successful.format(name=contact.name))
 
     def exit_app(self):
-        view.show_message(text.end_of_program)
-        view.input_date(text.press_any_key)
-        exit()
+        self.is_running = False
 
     def run(self):
         """
@@ -82,13 +105,16 @@ class MainController:
             self.delete_contact,
             self.exit_app
         ]
+        menu = text.main_menu_items.get('title', '')
+        options = text.main_menu_items.get('items', [])
 
-        is_running = True
-        while is_running:
+        while self.is_running:
             # Отрисовка меню
-            menu = text.main_menu_items.get('title', '')
-            options = text.main_menu_items.get('items', [])
             view.show_menu(menu, options)
             # Опции меню
             user_choice = view.get_user_menu_choice(options)
             menu_actions[user_choice - 1]()
+        else:
+            view.show_message(text.end_of_program)
+            view.input_date(text.press_any_key)
+            exit()
